Extract following/followers
