+++ src/App.css
@@ -5,6 +5,7 @@
 .App-logo {
   height: 8vmin;
   pointer-events: none;
+  margin: 6vmin 0;
 }
 
 .App-header {
