+++ src/components/follow.css
@@ -1,0 +1,3 @@
+span~span {
+    margin-left: 4px;
+}
